Move paths around to get the best of both worlds
Prepending PHP lets us switch PHP using `phps` but will override other
tools, for instance: curl, with PHP's version. In most cases we don't
want to be tied to PHP for these things.

Should you notice a tool is being taken over by PHP, symlink it in
$HOME/bin to be sure.

Evaluate this method after a while to see if many tools need to be
linked in this manner.

Also check if possible to add the last version of a directory for those
brewed applications. Symlinking the version to bin will inevitably lead
to hiccups.

Also important; stopped exporting the PATH on every line. This allegedly can
lead to some issues although I'm not sure what effect this could have
changing it at this point. If anything blows up, you'll know where to
look.